chore: data-display stories clean up
