+++ docs/components/Data Display/Collapse/Documentation.mdx
@@ -26,4 +26,4 @@ We use expand/collapse in [Table Filters](https://mparticle.github.io/aquarium/?
 | AntD | [Collapse Component](https://ant.design/components/collapse) |
 
 
-<Canvas meta={CollapseStories.default} />
+<Canvas of={CollapseStories.GhostCollapse} />
