+++ src/components/data-display/List/List.stories.tsx
@@ -1,5 +1,7 @@
 import { type Meta, type StoryObj } from '@storybook/react'
+import { Flex, Typography } from 'src/components'
 import { List } from 'src/components/data-display/List/List'
+import { ExampleStory } from 'src/utils/ExampleStory'
 
 const meta: Meta<typeof List> = {
   title: 'Components/Data Display/List',
@@ -8,26 +10,6 @@ const meta: Meta<typeof List> = {
   args: {
     dataSource: ['item1', 'item2', 'item3'],
     renderItem: item => <List.Item>{item as string}</List.Item>,
-    bordered: false,
-    footer: undefined,
-    grid: undefined,
-    header: undefined,
-    itemLayout: 'horizontal',
-    loading: false,
-    locale: { emptyText: 'No Data' },
-    pagination: false,
-    size: 'default',
-    split: true,
-  },
-  argTypes: {
-    itemLayout: {
-      control: 'select',
-      options: ['horizontal', 'vertical'],
-    },
-    size: {
-      control: 'select',
-      options: ['default', 'large', 'small'],
-    },
   },
 }
 export default meta
@@ -41,50 +23,48 @@ type Story = StoryObj<typeof List>
 
 export const Primary: Story = {}
 
-export const BorderedList: Story = {
+export const CustomHeader: Story = {
   args: {
-    bordered: true,
+    header: <Typography.Text strong>List Header</Typography.Text>,
+    dataSource: ['item1', 'item2', 'item3'],
   },
 }
 
-export const VerticalLayout: Story = {
-  args: {
-    itemLayout: 'vertical',
-  },
-}
+export const WithMetaItems: Story = {
+  render: () => {
+    const dataSource = [
+      {
+        title: 'item1',
+        description: 'description1',
+      },
+      {
+        title: 'item2',
+        description: 'description2',
+      },
+      {
+        title: 'item3',
+        description: 'description3',
+      },
+    ]
 
-export const LoadingIndicator: Story = {
-  args: {
-    loading: true,
-  },
-}
-
-export const CustomPagination: Story = {
-  args: {
-    pagination: { current: 1, total: 50, pageSize: 10 },
-  },
-}
-
-export const CustomGrid: Story = {
-  args: {
-    grid: { gutter: 16, column: 4 },
-  },
-}
-
-export const LargeSizeList: Story = {
-  args: {
-    size: 'large',
-  },
-}
-
-export const CustomFooter: Story = {
-  args: {
-    footer: <>Custom Footer</>,
-  },
-}
-
-export const NoSplitUnderListItem: Story = {
-  args: {
-    split: false,
+    return (
+      <ExampleStory title="A List that supports title and description.">
+        <List
+          dataSource={dataSource}
+          itemLayout="vertical"
+          renderItem={item => (
+            <List.Item key={item.title}>
+              <Flex gap="middle">
+                <List.Item.Meta
+                  title={<Typography.Text strong>{item.title}</Typography.Text>}
+                  description={item.description}
+                  className="u-margin-bottom-none"
+                />
+              </Flex>
+            </List.Item>
+          )}
+        />
+      </ExampleStory>
+    )
   },
 }
